refactor: gemini 리뷰 반영, 중복된 필터 포맷 로직을 유틸리티 함수로 추출(#159)
